refactor: extract icons in crepe
