+++ packages/crepe/src/feature/link-tooltip/index.ts
@@ -1,6 +1,6 @@
 import { configureLinkTooltip, linkTooltipConfig, linkTooltipPlugin } from '@milkdown/kit/component/link-tooltip'
 import type { DefineFeature } from '../shared'
-import { confirmIcon, deleteIcon, editIcon, linkIcon } from './consts'
+import { confirmIcon, editIcon, linkIcon, removeIcon } from '../../icons'
 
 export const defineFeature: DefineFeature = (editor) => {
   editor
@@ -10,7 +10,7 @@ export const defineFeature: DefineFeature = (editor) => {
         ...config,
         linkIcon: () => linkIcon,
         editButton: () => editIcon,
-        removeButton: () => deleteIcon,
+        removeButton: () => removeIcon,
         confirmButton: () => confirmIcon,
       }))
     })
