+++ packages/crepe/src/icons/bullet.ts
@@ -1,0 +1,14 @@
+import { html } from 'atomico'
+
+export const bulletIcon = html`
+  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
+    <g clip-path="url(#clip0_952_6527)">
+      <circle cx="12" cy="12" r="3"/>
+    </g>
+    <defs>
+      <clipPath id="clip0_952_6527">
+        <rect width="24" height="24"/>
+      </clipPath>
+    </defs>
+  </svg>
+`
