+++ packages/crepe/src/icons/index.ts
@@ -1,0 +1,33 @@
+export * from './alignCenter'
+export * from './alignLeft'
+export * from './alignRight'
+export * from './bold'
+export * from './bullet'
+export * from './bulletList'
+export * from './caption'
+export * from './checkBoxChecked'
+export * from './checkBoxUnchecked'
+export * from './clear'
+export * from './code'
+export * from './confirm'
+export * from './divider'
+export * from './dragHandle'
+export * from './edit'
+export * from './h1'
+export * from './h2'
+export * from './h3'
+export * from './h4'
+export * from './h5'
+export * from './h6'
+export * from './image'
+export * from './italic'
+export * from './link'
+export * from './menu'
+export * from './orderedList'
+export * from './plus'
+export * from './quote'
+export * from './remove'
+export * from './strikethrough'
+export * from './table'
+export * from './text'
+export * from './todoList'
